Backend Blueprints + Cleanup
